requirements.txt
openai>=1.6.0
anthropic>=0.8.0
langchain>=0.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.5.0
loguru>=0.7.2
networkx>=3.2.1
pandas>=2.0.0
numpy>=1.24.0
PySide6>=6.6.0
pyvis>=0.3.0
qt-material>=2.14